requirements.txt
yoyo-migrations
PyMySQL
sqlalchemy